Update patch set 1
Patch Set 1:

/neutron/model/src/main/yang/neutron.yang
Line 53:
Needs to be extracted and specialized type for individual uses of identifier. I think each domain has distinct identifiers and we certainly do not want to confuse port ID for a net ID.
Line 105:
this looks like yet another type of identifier
Line 108:
'address pairs' seems to imply a pair of leaves.
Line 111:
would this be a use of a networkUUID?
Line 117:
are these DHCP options? what is the expected string format?
Line 120:
what is the relationship between ID and opt_name?
Line 123:
a reference do definition would be nice.
Line 153:
what is the format of the string? is this subnetUUID?
Line 156:
inet:ip-address
Line 161:
are these identifiers?
Line 175:
definitely an enum
Line 185:
not really... new revision perhaps?
Line 189:
inet:ip-address? or can they actually be symbolic?
Line 203:
looks like an enumeration. if it needs to be externally extensinble, it should be an identity
Line 239:
if it is 'either IPv4 or IPv6', just use inet:ip-address

Patch-set: 1