Update patch set 6
Patch Set 6:

(11 comments)

Yes, it is a great improvement, but we should break it down further - we are conflating extensions and uuid is the atomic identifier for all base neutron objects.  Since it is not applied as part of an extension, prepending it with the type doesn't make sense to me.

Patch-set: 6
Label: Verified=0