Import-Package jaxb.rs version eclipse.persistence.version to avoid importing an older version.
Another core bundle loads version 2.5.0 ( I think a config bundle ) and if you later "feature:install" neutron it wires to this older version instead of the version 2.5.2 which neutron requires.   If you try to POST to the neutron REST with version 2.5.0 wired you get class not found exception.

Apparently ODL doesn't get this feature:install approach to building a controller, but my party needs this feature for our controller.

Change-Id: Ia8b97df585b0d96dcca18fd54cab4628d621a95d
Signed-off-by: jameshall03885 <ghall@brocade.com>