Bug 3072: Make NeutronRoutersNorthbound.addRouterInterface() idem-potent
With changes in Openstack's ml2 api [1], ODL will begin receiving
neutron northbound events for router interface create. Because of
that, we discovered a day 1 bug where neutron port associated with
the tourer interface will be already populated with the deviceId
and deviceOwner of the router interface. Prior to this gerrit,
the code would return ResourceConflictException() without
realizing that these two attributes had the proper values.

[1]: https://review.openstack.org/#/c/174600/

Change-Id: I69c22dc995d507111673389da110cd9e8db39570
Signed-off-by: Flavio Fernandes <ffernand@redhat.com>