Pay down technical debt: remove trailing comments
Squid includes a rule that Comments should not be located at the
end of lines of code to ensure readability. The main idea behind
this rule is that in order to be really readable, trailing comments
would have to be property written and formatted (correct alignment,
no interference with the visual structure of the code, not too long
to be visible) but most often, automatic code formatters would not
handle this correctly: the code would end up less readable.

Therefore, comments are far better placed on the previous empty
line of code, where they will always be visible and properly
formatted.

Change-Id: I5b1db07b0995d48508f7770a88ad807326b5a0e5
Signed-off-by: Ryan Moats <rmoats@us.ibm.com>