Update patch set 3
Patch Set 3: Code-Review-1

That thread is nice! Would be good to have a bug filed for it to track this work, and the commit message include a bit more. Am I being pedantic? Yes, I am. But in the interest of future folks looking at git history, it would be nice for them to not have to open up a web browser and/or bother you, Ryan, me, or Ed in IRC to understand what you meant here.

:)

Patch-set: 3
Reviewer: Gerrit User 254 <254@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1