Update patch set 1
Patch Set 1: Code-Review-1

> We can (and should do the deprecated decoration) on the I*Aware
 > classes
 > 
 > We can't make the changes to northbound-api classes until Boron

Hmm... if that is the case, how can an app (e.g. ovsdb-netvirt [1]) be able to switch to a native md-sal based? Are we expecting dummy-provider to be used in such cases? Otherwise, there will be no other
bundle with I*Aware instances, and no neutron events will be accepted.

 > 
 > Please split out the I*Aware changes as a separate patch

Ack; will do.

Patch-set: 1
Label: Code-Review=-1