Update patch set 8
Patch Set 8:

(8 comments)

This is important work consistent with the work in progress in the networking-bgpvpn Openstack project [1].

The major comment I would make is that we are discussing a change of our API (better do it now that the project is at its early phase) that this change will have to accomodate with: we are making network associations and router associations, API resources in their own right, as subresources of a BGPVPN object, so as to allow each association to have associated attributes, which is not possible if associated networks and routers are simple attributes of a bgpvpn object.  This will have to be accomodated in your Yang modelling and java object modelling as well.

This work really goes into the right direction and would deserve a +1, but we need to reflect that mismatch (which is due to us not having chosen the right model from day one).

Patch-set: 8