Bug 4527 - Security Group Rule doesn't allow all protocols
the API change to be made and that this patch is seeking a waive of
the API FREEZE.

Neutron Security Group Rule only accepts tcp, upd, icmp
and icmp6. Doesn't accept other protocols. This patch
changes the security rule protocol field to a union of
string and short to allow accepting other protocols.

Change-Id: I252b55807906103e3136c933547bbcac421b8090
Signed-off-by: Ritu Sood <ritu.sood@intel.com>