Update patch set 2
Patch Set 2:

What about usage of NeutronSubnet_HostRoutes? Ideally it should be replaced with something generic like NeutronRoutes but that might break code using it so went with safer option. I tried to avoid changing yang but it had to be changed. If we've raised a weather, want to try fix it proper way now, or let this be and we will take care of it in Boron as part of fixing all yangs?

Patch-set: 2
Label: Verified=0