Adding Yang Model for Openstack Host Configuration
Adding YANG module to define host configurations that is used
in specifying the physical node type (network node, compute node)
for host system. This information is used by Openstack Neutron
scheduler. For example for L3 centralized SNAT support this
will be used to inform Neutron which node is Network node.

Change-Id: I630e6d88f1b1beb6d5e442ccddf6a5d7137a5a51
Signed-off-by: Ritu Sood <ritu.sood@intel.com>