This patch enables the Quality of Service feature in the OpenDayLight Neutron. The implementation is done as follows 1. Defining the YANG model 2. Implementing of API 3. Testing the developed API
The QOS is achieved using either limiting the bandwidth or through
DSCP marking

Update about the patch.
- Changes in YANG model, SPI and the API as per the recent changes
in REST API format of neutworking-odl. Also fixed the changes as per comments.

Data model -
https://specs.openstack.org/openstack/neutron-specs/specs/liberty/qos-api-extension.html
This link has the data model changes as suggested by the OpenStack
Neutron for the QOS feature
API -
http://developer.openstack.org/api-ref-networking-v2-ext.html#qos-ext
This link has the API for the Openstack for the QOS feature

Change-Id: I7074dbb6d494b1267b66a925742e9812ef8a9304
Signed-off-by: pramod <pramod.raghavendra.jayathirth@intel.com>