Bug 5831 - build: check style error
This patch catches up for changes of odlparent checkstyle.
The checkset of 3b8a0b31dbe283fe37218dae2b997dd293283752 in odlparent
changed headerLocation of checkstyle from LISENCE.txt to
EPL-LICENSE.regexp.txt
https://maven.apache.org/plugins/maven-checkstyle-plugin/checkstyle-mojo.html#headerLocation
It causes the following error.
[ERROR] Failed to execute goal org.apache.maven.plugins:maven-checkstyle-plugin:2.16:check (default) on project project-neutron-parent: Failed during checkstyle execution: Failed during checkstyle configuration: unable to read /home/yamahata/openstack/odl/odl/upstream/neutron-maven-upload/parent/target/checkstyle-checker.xml - unable to parse configuration stream - Property ${checkstyle.header.file} has not been set -> [Help 1]

This patch removes the special configuration for neutron northbound so that
it uses odlparent config.

Change-Id: Ib499f2ce0b25a11c2dc73b716a69c78f057f70cf
Signed-off-by: Isaku Yamahata <isaku.yamahata@intel.com>