Update patch set 2
Patch Set 2:

Hi Brady, As isaku mentioned, this is to enable users to provide a SFC implementation/provider for the networking-sfc APIs. So these models are the ODL northbound representation for the networking-sfc API's. But i understand your question, that is the next item in my plate, to write a translation layer that can convert these networking-sfc yang model to the ODL SFC yang model, so that we can consume ODL-SFC project as one of the provider. I will share more details on it sometime next week with you/SFC team. I want to use this exercise to also determine the additional information that we need from networking-sfc APIs to translate it to ODL SFC. So lets discuss it on separate thread sometime next week, when i will have some more details.

Patch-set: 2
Label: Verified=0