Update patch set 1
Patch Set 1:

Hi Isaku,

I think the only logical solution would be to use ODL SFC as the backend, as writing a service provider from scratch would be reinventing the wheel.

An adapter layer will need to be written from networking-sfc to ODL SFC. Originally, I thought that layer would be in the networking-sfc project, which would then directly call the ODL SFC API. We should check with Tim Rozet if that is still the case. If it is, then this particular ODL API is not necessary.

If ODL does indeed need to implement this, then we need to figure out how to map from neurton-sfc to ODL SFC, which will most likely require pushing changes to the neutron-sfc project. I know Tim Rozet has already started this for the ODL SFC NSH configuration.

We should work out these detail before proceeding with this patch.

Patch-set: 1