Yang Model for Openstack Neutron Trunkport API's
This patch includes the yang model for the trunkport Openstack Neutron v2 extension.
The trunkport model represents the trunk ports and the subports configured in neutron.
Differences between the yang model and the neutron NB REST API are made because
several REST APIs in the neutron NB are RPCs, and we wish the keep the neutron NB
model simple to use.

References:
https://blueprints.launchpad.net/neutron/+spec/vlan-aware-vms
http://specs.openstack.org/openstack/neutron-specs/specs/newton/vlan-aware-vms.html
https://wiki.openstack.org/wiki/Neutron/TrunkPort

Change-Id: I0d0319bb9f2f0100c21446e2d82f26bec4ab8d55
Signed-off-by: Ariel <ariel.waizel@hpe.com>