Update patch set 3
Patch Set 3:

When sub-port is created/deleted it will also have uuid of trunk port it belongs to. So, we can add trunk-port uuid to sub-port object. It doesn't need to be an actual cross-reference, just uuid. As per API, trunk port can only be deleted once all sub-ports are deleted. So, by the time we need to act on a trunk-port, all subports would already be deleted and we won't need to refer all sub-ports for a given trunk port in NN code.

To send sub-ports as list in trunk port we will have to write the driver code to get all sub-ports associated with that trunk and add/delete on add/delete-sub-port in driver code.

Patch-set: 3
Label: Verified=0