Update patch set 3
Patch Set 3:

> When sub-port is created/deleted it will also have uuid of trunk
 > port it belongs to. So, we can add trunk-port uuid to sub-port
 > object. It doesn't need to be an actual cross-reference, just uuid.
 > As per API, trunk port can only be deleted once all sub-ports are
 > deleted. So, by the time we need to act on a trunk-port, all
 > subports would already be deleted and we won't need to refer all
 > sub-ports for a given trunk port in NN code.
 > 
 > To send sub-ports as list in trunk port we will have to write the
 > driver code to get all sub-ports associated with that trunk and
 > add/delete on add/delete-sub-port in driver code.

About trunk port deletion, the vlan aware vms spec says:

"The following deletion constraints exist:

Deletion of a trunk automatically deletes all of its subports.
Deletion of a (child) port referred by a subport is forbidden. The subport must be deleted first.
Deletion of a (parent) port referred by a trunk is forbidden. The trunk must be deleted first."

So deletion of the trunk port is possible while it still has sub ports. The constraint on deletion is on deleting the "legacy" port before deleting the trunk port/ sub port.

Patch-set: 3