Update patch set 1
Patch Set 1:

The Karaf 3 features have a feature called odl-neutron-service, which is basically just pulling together odl-neutron-spi, odl-neutron-northbound-spi, and odl-neutron-transcriber. I see it is missing from this patch. Are you going to provide it in an updated patch set, or is it deprecated, and projects should just use the features they really need? We are currently using odl-neutron-service in lispflowmapping.

Patch-set: 1
Label: Verified=0