Update patch set 5
Patch Set 5:

> Looking at the log, integration test isn't run.
 > https://logs.opendaylight.org/releng/jenkins092/neutron-distribution-check-nitrogen/6/console.log.gz
 > [INFO] --- maven-failsafe-plugin:2.18.1:verify (integration-tests)
 > @ integration-test ---
 > [INFO] Tests are skipped.
The tests are 
In the distribution check job it always show Tests are skipped. This is same for boron as well. 

 > Looking at the log, integration test isn't run.
 > https://logs.opendaylight.org/releng/jenkins092/neutron-distribution-check-nitrogen/6/console.log.gz
 > [INFO] --- maven-failsafe-plugin:2.18.1:verify (integration-tests)
 > @ integration-test ---
 > [INFO] Tests are skipped.

In neutron-verify-nitrogen-mvn33-openjdk8 job it shows that Tests are executed. In distribution check jobs it shows as Tests skipped. This is same for boron as well.

Patch-set: 5
Label: Verified=0