Update patch set 1
Patch Set 1:

That's the problem... many projects use the utils rather than duplicating them. Worked fine since it really just used southbound, but that last patch pulled in the library. Creating a feature would be the best going forward or as Vishal mentioned on the thread to include the utils in the southbound-api which doesn't seem like so much of a stretch.

Or would it be cleaner in this patch to just include the odl-ovsdb-southbound-impl along with the utils? That is what the netvirt and genius are doing and works fine.

Patch-set: 1
Reviewer: Gerrit User 773 <773@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=0