Neutron Model update for Status field
Breaking out the status field from the other model
update patch for ease of review and merge process.
https://git.opendaylight.org/gerrit/#/c/51534/

This patch includes Neutron API/model changes for
Status field to be moved to operational datastore
from configuration as per discussion on this thread:
https://lists.opendaylight.org/pipermail/netvirt-dev/2017-February/003141.html

NN reads and writes to configuration datastore only.

Change-Id: I9a7d23e878507edfd675c3fb17016fe15c7ccd08
Signed-off-by: Ritu Sood <ritu.sood@intel.com>