Update patch set 10
Patch Set 10: Code-Review-1

It might be smarter to temporarily leave the old port status in place in the yang model to ease with the transition. Currently this change will require changes in netvirt and in networking-odl. If you leave the old port status field in the model until those changes go through, nothing will break.

Patch-set: 10
Label: Code-Review=-1