Update patch set 4
Patch Set 4:

Yes! I've tried integrating your commit with the netvirt commit I'm working on. it's a bit of a challenge because after parsing the profile to NeutronPort (works just fine) we pass the data to the Port object where the profile is defined as a string. I'll upload the changes to a different branch once i'll get it to work.

Patch-set: 4
Reviewer: Gerrit User 7595 <7595@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0