Move jaxb.properties from wrong package to the correct one
although the fact that everything works anyway clearly means that this
actually is not really currently needed.  I believe, based on
NEUTRON-154, that this is because the MOXyJsonProvider which we
configure in NeutronNorthboundRSApplication probably already does the
equivalent programmatically in code to what this jaxb.properties does
for JAXB.

However, instead of just deleting it, I do think keeping it here but in
the right place has value for being able to write unit tests for the
JAXB annotated API model binding classes in the future, if ever, e.g. to
unit test their JSON (de)serialization outside of JAX RS (and thus
NeutronNorthboundRSApplication) via standard JAXB APIs from such future
tests.

Change-Id: I98d001b862e8cc69529112a330b5195698889b5c
Signed-off-by: Michael Vorburger <vorburger@redhat.com>