Simplify odl-neutron-northbound-api POM
all of the removd dependencies are not required, because they are
already transitively inherited from northbound-api, and duplicating them
here adds no value, and only confusion (especially if we change them in
northbound-api; as in NEUTRON-154 or maybe later in NEUTRON-124 or if
ever switching from Moxy to GSON or something).

This is safe and should not break anything, because if it did then the
ITNeutronE2E (which still passsed locally) would detect any major
regressions.

Change-Id: I6b232321c2e73e55f050a74b277ac3894977ed59
Signed-off-by: Michael Vorburger <vorburger@redhat.com>