Simplify odl-neutron-northbound-api POM
all of the dependencies removed are not required, because they already
are in and transitively inherited from northbound-api; so duplicating
them here adds no value, and only confusion (especially if we change
them in northbound-api; as in NEUTRON-154 or maybe later in NEUTRON-124
or if ever switching from Moxy to GSON or something).

This is safe and should not break anything, because if it did, then the
ITNeutronE2E would detect any major regressions (and it still passes
with this change).

Change-Id: I6b232321c2e73e55f050a74b277ac3894977ed59
Signed-off-by: Michael Vorburger <vorburger@redhat.com>