replace Jackson specific @JsonIgnore annotation with JAXB @XmlTransient
on NeutronObject's setTenantID(Uuid tenantID), but note
setTenantID(String tenantID), although with @XmlElement(name =
"tenant_id") on field String tenantID and
@XmlAccessorType(XmlAccessType.NONE) annotating the setter was probably
pointless anyway,

and remove dependency to jackson-annotations,

because neutron (apparently, so learnt in NEUTRON-154!) uses Moxy and
not Jackson, at least clearly now it does, perhaps in the past it once
used Jackson, so this annotation was probably already always ignored
anyway.

This is safe and should not break anything, because if it did, then the
ITNeutronE2E would detect any major regressions (and it still passes
with this change).

found while rummaging around neutron code related to NEUTRON-154

Change-Id: I6b815cdade1a7e0877c456520f7f0f5b9215057c
Signed-off-by: Michael Vorburger <vorburger@redhat.com>