Update patch set 1
Patch Set 1:

(3 comments)

Thank you so much!

Later soon we'll switch to use spotbugs instead of findbugs. So you may want to fix it against spotbugs.
Maybe this patch can be split, so that we can merge them independently without carrying big patch. but it's quite up to you.

If you'd like, we can merge this patch and address issues with followup patches. or you can respin this patch. which way do you prefer?

Patch-set: 1