Update patch set 1
Patch Set 1:

(3 comments)

> (3 comments)
 > 
 > Thank you so much!
 > 
 > Later soon we'll switch to use spotbugs instead of findbugs. So you
 > may want to fix it against spotbugs.

I haven't tried spotbugs but my understanding it is a re-incarnation of findbugs, ie the original findbugs project died out and was taken over by others and renamed. It's possible we may hit more violations with spotbugs but those can be addressed then. 

 > Maybe this patch can be split, so that we can merge them
 > independently without carrying big patch. but it's quite up to you.

You mean split this up into a patch per bundle. I should probably have done that but I didn't think there would be this many violations and then I got on a roll.

 > 
 > If you'd like, we can merge this patch and address issues with
 > followup patches. or you can respin this patch. which way do you
 > prefer?

Most of the changes are cosmetic so I don't anticipate runtime issues but is there anyway this patch can be tested to ensure that? I'm not familiar with neutron stuff. If there is a possibility for breakage, we're past M4 so we should probably hold off on merging this until after the Oxygen branch is cut.

Patch-set: 1