Update patch set 1
Patch Set 1:

I see Isaku's point but the changes are really minor and look to be (mostly) from Eclipse save actions like adding @Override. This happens to me too and I don't see it until I submit the changes. It would be a bit of a pain to disable the save actions and remove those changes. Frankly - not worth the effort to do it.

Patch-set: 1