replace web.xml with programmatic registration
this is just the first step; once this is in,
we can replace much of the static etc. and directly
wire the web components and the OSGi beans and services.

The blueprint.xml is needed because of the weird AAA-170
issue with the pax-cdi-api @OsgiService annotation.

Using change Ia3a16df71384610a75acf3d28205c973c554d477
in AAA, we directly use the WebContextSecurer instead of
relying on aaa-shiro in code because that makes Neutron
code de-coupled from AAA (as it should be) and prevents
the KarafIniWebEnvironment dead-lock, see AAA-171.

Change-Id: I1dc13069b7e0ec456941f523e9dd4d560841f3e6
Signed-off-by: Michael Vorburger <vorburger@redhat.com>