replace web.xml with programmatic registration
this is just the first step; once this is in,
we can replace much of the static etc. and directly
wire the web components and the OSGi beans and services.

The blueprint.xml is needed because there is some weird
issue with the pax-cdi-api @OsgiService annotation IFF
(only) a Service has a ServiceFactory like WebServer.

Change-Id: I1d141156a30ef96de6d21060779d21e426532a3c
Signed-off-by: Michael Vorburger <vorburger@redhat.com>