Add Karaf build profile
Project local Karaf distros are handy for devs to test their work
however is unneeded by autorelease builds and should not be released
as part of the Simultanious Release. Add a profile that is active by
default so that default behaviour is unchanged however allows the
autorelease project to disable building this module.

Issue: RELENG-86
Change-Id: I09f95138ae5b5eb073d8556d36a2b47ec83a912d
Signed-off-by: Daniel Farrell <dfarrell@redhat.com>