Update patch set 3
Patch Set 3: Code-Review-1

I'll help to review this, but gimme 1h to run this with a revert of ab548fe1a7018d5592f315f59589c5a1da59059c, so through the ITNeutronE2E (which we commented out in that commit), just to be sure... Temporary -1 just because of that.  I'll also rebase it.

Patch-set: 3
Label: Code-Review=-1