Update patch set 5
Patch Set 5:

Thanks for figuring out ClassPathHell issues. I had commented it out  temporarily so as not to block the jersey 2 work and revisit it later.

So the netconf and aaa patches need each other but this ONE seems to be OK on its own. Still do you want to merge all 3 at/around the same time to be safe?

Patch-set: 5