Update patch set 1
Patch Set 1:

> > > (1 comment)
 > >
 > > We moved from replacing enums in yang with identity during early
 > > Lithium timeframe. Use of enums is not recommended as it provides
 > > no validation. Generated code is all strings, so any value can be
 > > written into datastore, valid or not. You can look at any of
 > > existing yangs in Neutron/Netvirt/Genius and everywhere identity
 > is
 > > used for enum use cases.
 > 
 > There are lot of enums in netvirt code - at least for
 > configurations.
 > I did check the acl-service-config.yang where acl behavior is
 > defined by "allow" and "deny" values as enums. And the generated
 > code is java enum, it won't allow any random values.
 > Probably during lithium timeframe it might have allowed, might not
 > be the case now.

But this is not configuration yang, this is for rest. Different use cases. config.xml parameters are safe to use enum coz they're populated and validated differently from these API yangs. It was not a case of what enum can or not do, it was recommendation from MDSAL/yangtools team to stop using enums and use identity instead.

Patch-set: 1