Update patch set 2
Patch Set 2:

(1 comment)

> lets wait for TSC-115 to make sure this doesn't break other
 > projects..

multipatch on TSC-115 passed, so this is "just" a normal Neutron change, not really a weather item breaking other projects, it seems.

Keeping -1 for a very minor style point I noticed; then +1 from me.

 > Looks like NeutronUtils.java is in generated directory. If
 > possible, move it somewhere outside.

Perhaps that could be done in separate new follow-up change? Because that NeutronUtils is currently already there in that package (not directory) which typically has generated code, so it's not like change introduces that - this change seems to fix NEUTRON-163, so doesn't have to clean up something that's already there.

What WOULD be really nice would be TDD for NEUTRON-163, so for this change to include a test which failed before this and passes with this...

Patch-set: 2