Update patch set 6
Patch Set 6:

So the failure was due to the bundle failing to resolve due to missing service imports due to the generated Import-Service directive in the manifest.  The  Import-Service requires the service be present when the bundle is resolved which isn't the case b/c the transcriber bundle is also being installed - if the transcriber bundle was already installed and running then it would resolve. 

So the question is why is the Import-Service directive added....  strangely enough it isn't added if the blueprint XML file exists - so leaving an empty blueprint causes the Import-Service to be omitted and SFT succeeds. I haven't looked into why or if there's a way to disable the generation of Import-Service.

Patch-set: 6