propagate datastore exceptions all the way to northbound
Instead of logging but ultimately ignoring them, as they currently are.

Stumbled across this wile working on (not directly related) NEUTRON-158.

This will also fix the problem raised in NEUTRON-157: We are,
intentionally, changing the OptimisticLockFailedException WARN log to a
DEBUG only.  If it still didn't go through after RETRY_MAX (currently 2)
then it will be rethrown, and that, with this, will go back up to the
driver.

JIRA: NEUTRON-157
Change-Id: I10b7dea0b5698db40ee7cfaa63593fb6b3c573c7
Signed-off-by: Michael Vorburger <vorburger@redhat.com>