{WIP} improve error logs due to IAE to include full input
This requires, and adds, a working to String for all *Request classes,
which is anyway useful for future logging of requests elsewhere.

TODO all *Request classes, not just the 3 LoadBalancer related ones,
must extends NeutronRequest instead of implements INeutronRequest.

This introduction of a NeutronRequest (an ABC instead of an interface)
also allows a follow-up change to get rid of the stupid reflection in
INeutronRequest by having a (generic) "singleton" and a "bulkRequest"
field in NeutronRequest instead of in each *Request class.

Change-Id: I9769a3825632dd2bd2f9e2875b91d06a4f781616
Signed-off-by: Michael Vorburger <vorburger@redhat.com>