Restore duplicate check
With the recent odlparent and yangtools bumps in Fluorine, the
handling of jaxb-api and activation changed; for features, this is
dealt with by excluding the relevant artifacts from generated
features, but other tests end up with duplicates since the artifacts
include classes which are also present in the JVM’s rt.jar.

This patch excludes jaxb-api and activation, and restores the
duplicate check.

JIRA: INFRAUTILS-59
Change-Id: Ifa8f0fd86c8b65655cc50261f198d3deb2411499
Signed-off-by: Stephen Kitt <skitt@redhat.com>