Update patch set 10
Patch Set 10:

Inline...

 > > I have a number of small feedbacks (see inline), and a more
 > general
 > > larger feedback... I don't understand at all what this actually
 > > solves, and how! ;)  Is there a JIRA (or RH BZ) with more
 > > background than the Commit Message?
 > 
 > FTR: apparently this is related to / intended to fix
 > https://jira.opendaylight.org/browse/NETVIRT-1460 (and also has
 > something to do with https://jira.opendaylight.org/browse/OPNFLWPLUG-1039
 > ?), but I'm not clear about the details.

The details are as follows. In order for a websocket connection to work it must be proceeded by two rest calls that complete the registration. The issue we were seeing is that due to a switch to round robin balancing there were cases where a networking_odl would attempt to connect to an ODL that had not received the registration calls. There was a whole discussion about how to solve this and what seemed most solid is to have each ODL pre-register itself and not rely on the REST calls from the client.

 > 
 > >
 > > At first look, making Neutron dependant on RestConf (which, as
 > far
 > > as I can tell, it currently is not, from a code perspective), 

Just to be clear on the utility of this, we've done experiments with equivalent, hacky approaches and this change *greatly* improves our clustered CI. I think it's worth the dependency.

and
 > > making Neutron, at boot, invoke a RestConf WebSocket URL via HTTP
 > > on localhost seems... a curious approach.  

This does not invoke via HTTP...

Whatever this is for,
 > > shouldn't there be an internal API for what you need to do in
 > > RESTCONF?

I think it would have been nicer if there was an actual API for this as well but JSONRestconfService already existed. If I had designed it I would have created an actual Java API (not a Java API that takes json as input) then wrapped that in the http/rest API...but that is not what exists and I do not think it's worth creating it now.

Patch-set: 10
Reviewer: Gerrit User 5357 <5357@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0