Update patch set 5
Patch Set 5:

Mr @Vorburger,

So, based on your comments here and our IRC discussion it seems that we agree that retrying is the right approach for stable/oxygen. I've uploaded a new patch set where the retries happen in the constructor and not in a timer thread. Regarding your comment about why I print the exception using just the message and not the full stack trace, that's because that exception is no longer unexpected and I don't want to give the wrong impression in the logs.

Patch-set: 5
Reviewer: Gerrit User 5357 <5357@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0