Update patch set 2
Patch Set 2:

I am not convinced that blueprint should fail if this fails. It's fuzzy to begin with, retrying. IMHO, BP should fail only if the system is unusable. If this fails the system may not be unusable, registration can still happen via REST or something. I'm not convinced either way.

Patch-set: 2